Change API request and add Authentication check for PostController, CommentController
|> UserController: "Update", "Delete" now don't need request userId
|> PostController: "Create", "Update", "Delete" now don't need request userId
|> CommentController: "Create", "Update", "Delete" now don't need request
|> LikeController: "Toggle Like" don't need userId

+++ src/main/java/com/project/social_media/controllers/CommentController.java
@@ -2,24 +2,34 @@ package com.project.social_media.controllers;
 
 import com.project.social_media.application.DTO.CommentDTO;
 import com.project.social_media.application.Service.CommentService;
+import com.project.social_media.application.Service_Interface.IAuthenticationService;
+import com.project.social_media.application.Service_Interface.ICommentService;
+import com.project.social_media.application.Service_Interface.IPostService;
+import com.project.social_media.application.Service_Interface.IUserService;
 import com.project.social_media.controllers.ApiResponse.ApiResponse;
 import com.project.social_media.controllers.Request.Comment.CommentCreateRequest;
 import com.project.social_media.controllers.Request.Comment.CommentUpdateRequest;
 import com.project.social_media.domain.Model.Comment;
+import com.project.social_media.domain.Model.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
-import static org.springframework.http.HttpStatus.INTERNAL_SERVER_ERROR;
+import static org.springframework.http.HttpStatus.*;
 
 @Controller
 @RequiredArgsConstructor
 @RequestMapping("${api.prefix}/comments")
 public class CommentController {
-    private final CommentService commentService;
+    private final ICommentService commentService;
+    private final IUserService  userService;
+    private final IPostService postService;
+    private final IAuthenticationService authenticationService;
 
     @GetMapping("/all")
     public ResponseEntity<ApiResponse> getAllComment() {
@@ -34,8 +44,7 @@ public class CommentController {
     }
 
     @GetMapping("/all/post")
-    public ResponseEntity<ApiResponse> getAllCommentByPostId(
-            @RequestParam(required = false) Long postId) {
+    public ResponseEntity<ApiResponse> getAllCommentByPostId(@RequestParam(required = false) Long postId) {
         try {
             List<Comment> commentList = commentService.getAllCommentsByPostId(postId);
             List<CommentDTO> commentDTOList = commentService.convertToDTOList(commentList);
@@ -47,8 +56,7 @@ public class CommentController {
     }
 
     @GetMapping("/all/user")
-    public ResponseEntity<ApiResponse> getAllCommentByUserId(
-            @RequestParam(required = false) Long userId) {
+    public ResponseEntity<ApiResponse> getAllCommentByUserId(@RequestParam(required = false) Long userId) {
         try {
             List<Comment> commentList = commentService.getAllCommentsByUserId(userId);
             List<CommentDTO> commentDTOList = commentService.convertToDTOList(commentList);
@@ -61,11 +69,16 @@ public class CommentController {
 
     @PostMapping("/create")
     public ResponseEntity<ApiResponse> createComment(
-            @RequestParam Long userId,
             @RequestParam Long postId,
             @ModelAttribute CommentCreateRequest request) {
         try {
-            Comment newComment = commentService.createComment(userId, postId, request);
+            // Get authenticate user
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            authenticationService.authenticationCheck(authentication);
+            User authUser = userService.getUserByUsername(authentication.getName());
+
+            // Create
+            Comment newComment = commentService.createComment(authUser.getUserId(), postId, request);
             CommentDTO newCommentDTO = commentService.convertToDTO(newComment);
             return ResponseEntity.ok(new ApiResponse("Success", newCommentDTO));
         } catch (Exception e) {
@@ -76,13 +89,23 @@ public class CommentController {
 
     @PutMapping("/comment/{commentId}/update")
     public ResponseEntity<ApiResponse> updateComment(
-            @RequestParam Long userId,
             @RequestParam Long postId,
             @PathVariable Long commentId,
             @ModelAttribute CommentUpdateRequest request) {
-
         try {
-            Comment updatedComment = commentService.updateComment(userId, postId, commentId, request);
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            authenticationService.authenticationCheck(authentication);
+            User authUser = userService.getUserByUsername(authentication.getName());
+            Comment existingComment = commentService.getCommentById(commentId);
+
+            // Authentication Check
+            if (!authUser.getUserId().equals(existingComment.getUser().getUserId())) {
+                return ResponseEntity.status(UNAUTHORIZED)
+                        .body(new ApiResponse("Invalid Permission", null));
+            }
+
+            // Update
+            Comment updatedComment = commentService.updateComment(authUser.getUserId(), postId, commentId, request);
             CommentDTO updatedCommentDTO = commentService.convertToDTO(updatedComment);
             return ResponseEntity.ok(new ApiResponse("Success", updatedCommentDTO));
         } catch (Exception e) {
@@ -93,11 +116,20 @@ public class CommentController {
 
 
     @DeleteMapping("/comment/delete")
-    public ResponseEntity<ApiResponse> deleteComment(
-            @RequestParam Long commentId
-    ) {
-
+    public ResponseEntity<ApiResponse> deleteComment(@RequestParam Long commentId) {
         try {
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            authenticationService.authenticationCheck(authentication);
+            User user = userService.getUserByUsername(authentication.getName());
+            Comment existingComment = commentService.getCommentById(commentId);
+
+            // Authentication Check
+            if (!user.getUserId().equals(existingComment.getUser().getUserId())) {
+                return ResponseEntity.status(UNAUTHORIZED)
+                        .body(new ApiResponse("Invalid Permission", null));
+            }
+
+            // Delete
             commentService.deleteCommentById(commentId);
             return ResponseEntity.ok(new ApiResponse("Success", null));
         } catch (Exception e) {
